pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "bot"
version = "0.1.0"
[tool.poetry.dependencies]
Flask = "^1.1"
airtable-python-wrapper = "^0.13.0"
"discord.py" = "^1.3"
python = "^3.8"
[tool.poetry.dev-dependencies]
